requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
openai>=1.0.0
langchain>=0.1.0
langchain-openai
langchain-qdrant
langchain-core
qdrant-client
redis
bleach
fastapi-limiter
chainlit
tenacity
requests
beautifulsoup4
lxml
langfuse>=2.33.0